Create separate file to generate html.
